add substructure match method for fingerprints; also fix exact match method for fingerprints
