+++ src/analysis/mod.rs
@@ -1,2 +1,2 @@
 pub mod compound_processing;
-pub mod structure_search;
+pub mod structure_matching;
